+++ js/data.js
@@ -1,6 +1,8 @@
 import {getRandomIntInclusive, getRandomFloat} from './util.js';
 
-
+/*
+Переменне для генерации объявления
+*/
 const AVATAR_PIC_NUMBERS = 8;
 const MAXIMUM_PER_DAY_PRICE = 1000;
 const MAXIMUM_ROMS = 5;
@@ -17,6 +19,25 @@ const LAT_MINIMUM = 35.65000;
 const LAT_MAXIMUM = 35.70000;
 const LNG_MINIMUM = 139.70000;
 const LNG_MAXIMUM = 139.80000;
+
+/*
+Переменные для отрисовки объявления
+*/
+const OFFER_TEMPLATE = document.querySelector('#card').content; // находим содержимое шаблона
+const OFFER_CARD = OFFER_TEMPLATE.querySelector('.popup'); // находим объявление 
+const FRAGMENT = document.createDocumentFragment();
+const MAP_CANVAS = document.querySelector('#map-canvas');
+const OFFER_ELEMENT = OFFER_CARD.cloneNode(true);
+const TEMPLATE_TITLE = OFFER_ELEMENT.querySelector('.popup__title');
+const TEMPLATE_ADDRESS = OFFER_ELEMENT.querySelector('.popup__text--address');
+const TEMPLATE_PRICE = OFFER_ELEMENT.querySelector('.popup__text--price');
+const TEMPLATE_TYPE = OFFER_ELEMENT.querySelector('.popup__type');
+const TEMPLATE_GUESTS = OFFER_ELEMENT.querySelector('.popup__text--capacity');
+const TEMPLATE_TIME = OFFER_ELEMENT.querySelector('.popup__text--time');
+const TEMPLATE_FEATURES = OFFER_ELEMENT.querySelector('.popup__features');
+const TEMPLATE_DESCRIPTION = OFFER_ELEMENT.querySelector('.popup__description');
+const TEMPLATE_PHOTO = OFFER_ELEMENT.querySelector('.popup__photo');
+const TEMPLATE_AVATAR = OFFER_ELEMENT.querySelector('.popup__avatar');
 
 /**
 * Генерирует массив особенностей объекта проживания
@@ -101,4 +122,28 @@ const createOffers = () => {
   return tray;
 };
 
+const ADVERTISEMENTS = createOffers(); // формируем данные для объявлений
+
+/*
+Переменные для заполнения объявления
+*/
+const OFFER_TITLE = ADVERTISEMENTS[0].offer.title;
+const OFFER_ADDRESS = ADVERTISEMENTS[0].offer.address;
+const OFFER_PRICE = ADVERTISEMENTS[0].offer.price + ' ₽/ночь';
+const OFFER_TYPE = ADVERTISEMENTS[0].offer.type;
+const OFFER_ROOMS = ADVERTISEMENTS[0].offer.rooms;
+const OFFER_GUESTS = ADVERTISEMENTS[0].offer.guests;
+const OFFER_CHECKIN = ADVERTISEMENTS[0].offer.checkin;
+const OFFER_CHECKOUT = ADVERTISEMENTS[0].offer.checkout;
+const OFFER_FEATURES = ADVERTISEMENTS[0].offer.features;
+const OFFER_DESCRIPTION = ADVERTISEMENTS[0].offer.description;
+const OFFER_PHOTOS = ADVERTISEMENTS[0].offer.photos[0];
+const AVATAR = ADVERTISEMENTS[0].author.avatar;
+
+
 export {createAuthor, createLocation,createOffer,createOffers};
+export {OFFER_TEMPLATE,OFFER_CARD,FRAGMENT,MAP_CANVAS,ADVERTISEMENTS,
+  OFFER_ELEMENT,TEMPLATE_TITLE,TEMPLATE_ADDRESS, TEMPLATE_PRICE,TEMPLATE_TYPE,TEMPLATE_GUESTS,
+TEMPLATE_TIME,TEMPLATE_FEATURES,TEMPLATE_DESCRIPTION,TEMPLATE_PHOTO,TEMPLATE_AVATAR,OFFER_TITLE,OFFER_ADDRESS,
+OFFER_PRICE,OFFER_TYPE,OFFER_ROOMS,OFFER_GUESTS,OFFER_CHECKIN,OFFER_CHECKOUT,OFFER_FEATURES,
+OFFER_DESCRIPTION,OFFER_PHOTOS,AVATAR};
