+++ js/main.js
@@ -1,5 +1,4 @@
-import {createOffers} from './data.js';
-import {OFFER_TEMPLATE} from './offers.js';
+import { offerDraw } from './offers.js';
 
-// eslint-disable-next-line no-console
 
+offerDraw();
